[mobile] refactor: Picker関連の関数などをuse-casesに移動
